Added page to show matches

+++ MatchmakingPlatform/MatchmakingPlatform/BL/Preference.cs
@@ -7,11 +7,11 @@ using System.Threading.Tasks;
 namespace MatchmakingPlatform{
     public class Preference{
         public string Pref{get;set; } 
-        public float Value { get;set; }
+        public int Value { get;set; }
         public string Condition { get;set; }
 
 
-        public Preference(string Pref,float Value,string Condition)
+        public Preference(string Pref,int Value,string Condition)
         {
             this.Pref = Pref;
             this.Value = Value;
